+++ kompendium-core/src/main/kotlin/io/bkbn/kompendium/core/util/Helpers.kt
@@ -18,7 +18,7 @@ object Helpers {
   val UNIT_TYPE by lazy { Unit::class.createType() }
 
   /**
-   * Higher order function that takes a map of names to objects and will log their state ahead of function invocation
+   * Higher order function that takes a map of names to object and will log their state ahead of function invocation
    * along with the result of the function invocation
    */
   fun <T> logged(functionName: String, entities: Map<String, Any>, block: () -> T): T {