+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/FormattedSchema.kt
@@ -4,5 +4,11 @@ data class FormattedSchema(
   val format: String,
   override val type: String,
   override val default: Any? = null,
-  override val description: String? = null
+  override val description: String? = null,
+  // Constraints
+  val minimum: Int? = null,
+  val maximum: Int? = null,
+  val exclusiveMinimum: Boolean? = null,
+  val exclusiveMaximum: Boolean? = null,
+  val multipleOf: Int? = null,
 ) : TypedSchema